where can signalr be used, and useful? wire up a baseline
