+++ Workflow/Api/ActionsController.cs
@@ -6,11 +6,13 @@ using System.Net;
 using System.Reflection;
 using System.Web.Http;
 using log4net;
+using Microsoft.AspNet.SignalR;
 using Umbraco.Core.Models.Membership;
 using Umbraco.Web;
 using Umbraco.Web.WebApi;
 using Workflow.Models;
 using Workflow.Helpers;
+using Workflow.Notifications;
 using Workflow.Processes;
 using Workflow.Services;
 using Workflow.Services.Interfaces;
@@ -27,6 +29,7 @@ namespace Workflow.Api
 
         private readonly IInstancesService _instancesService;
         private readonly ITasksService _tasksService;
+        private readonly IHubContext _hubContext;
 
         private readonly Utility _utility;
 
@@ -36,6 +39,7 @@ namespace Workflow.Api
             _tasksService = new TasksService();
 
             _utility = new Utility();
+            _hubContext = GlobalHost.ConnectionManager.GetHubContext<PlumberHub>();
         }
 
         public ActionsController(UmbracoContext umbracoContext) : base(umbracoContext)
@@ -44,6 +48,7 @@ namespace Workflow.Api
             _tasksService = new TasksService();
 
             _utility = new Utility();
+            _hubContext = GlobalHost.ConnectionManager.GetHubContext<PlumberHub>();
         }
 
         /// <summary>
@@ -89,6 +94,10 @@ namespace Workflow.Api
                 }
 
                 Log.Info(msg);
+
+                _hubContext.Clients.All.WorkflowStarted(
+                    _instancesService.ConvertToWorkflowInstanceList(new List<WorkflowInstancePoco> { instance })
+                        .FirstOrDefault());
 
                 return Json(new
                 {
@@ -158,6 +167,10 @@ namespace Workflow.Api
 
                 Log.Info(logMsg);
 
+                _hubContext.Clients.All.TaskApproved(
+                    _instancesService.ConvertToWorkflowInstanceList(new List<WorkflowInstancePoco> { instance })
+                        .FirstOrDefault());
+
                 return Json(new
                 {
                     message = msg,
@@ -199,6 +212,10 @@ namespace Workflow.Api
 
                 Log.Info($"{instance.TypeDescription} request for {instance.Node.Name} [{instance.NodeId}] was rejected by {currentUser.Name}");
 
+                _hubContext.Clients.All.TaskRejected(
+                    _instancesService.ConvertToWorkflowInstanceList(new List<WorkflowInstancePoco> { instance })
+                        .FirstOrDefault());
+
                 return Json(new
                 {
                     message = instance.TypeDescription + " request has been rejected.",
@@ -239,6 +256,10 @@ namespace Workflow.Api
 
                 Log.Info($"{instance.TypeDescription} request for {instance.Node.Name} [{instance.NodeId}] was cancelled by {currentUser.Name}");
 
+                _hubContext.Clients.All.TaskCancelled(
+                    _instancesService.ConvertToWorkflowInstanceList(new List<WorkflowInstancePoco> { instance })
+                        .FirstOrDefault());
+
                 return Json(new
                 {
                     status = 200,
@@ -277,6 +298,10 @@ namespace Workflow.Api
                 );
 
                 Log.Info($"{instance.TypeDescription} request for {instance.Node.Name} [{instance.NodeId}] was resubmitted by {currentUser.Name}");
+
+                _hubContext.Clients.All.TaskResubmitted(
+                    _instancesService.ConvertToWorkflowInstanceList(new List<WorkflowInstancePoco> { instance })
+                        .FirstOrDefault());
 
                 return Json(new
                 {
