+++ Workflow/Notifications/PlumberHub.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNet.SignalR;
+
+namespace Workflow.Notifications
+{
+    public class PlumberHub : Hub
+    {
+
+    }
+}
