+++ Workflow/Processes/DocumentUnpublishProcess.cs
@@ -3,11 +3,11 @@ using System;
 using System.Linq;
 using Umbraco.Core;
 using Umbraco.Core.Models;
-using Umbraco.Core.Persistence;
 using Umbraco.Core.Services;
 using Workflow.Events.Args;
 using Workflow.Helpers;
 using Workflow.Models;
+using Workflow.Notifications;
 using Workflow.Services;
 using Workflow.Services.Interfaces;
 
@@ -22,7 +22,7 @@ namespace Workflow.Processes
         private static IContentService _contentService;
         private static IInstancesService _instancesService;
 
-        private readonly Notifications _notifications;
+        private readonly Emailer _emailer;
         private readonly Utility _utility;
 
         private static string _nodeName;
@@ -33,17 +33,17 @@ namespace Workflow.Processes
             : this(
                 ApplicationContext.Current.Services.ContentService,
                 new InstancesService(),
-                new Notifications(),
+                new Emailer(),
                 new Utility()
             )
         {
         }
 
-        private DocumentUnpublishProcess(IContentService contentService, IInstancesService instancesService, Notifications notifications, Utility utility)
+        private DocumentUnpublishProcess(IContentService contentService, IInstancesService instancesService, Emailer emailer, Utility utility)
         {
             _contentService = contentService;
             _instancesService = instancesService;
-            _notifications = notifications;
+            _emailer = emailer;
             _utility = utility;
 
             Type = WorkflowType.Unpublish;
@@ -110,7 +110,7 @@ namespace Workflow.Processes
 
             if (success)
             {
-                _notifications.Send(Instance, EmailType.ApprovedAndCompleted);
+                _emailer.Send(Instance, EmailType.ApprovedAndCompleted);
                 Log.Info("Successfully unpublished page " + Instance.Node.Name);
 
                 Completed?.Invoke(this, new InstanceEventArgs(Instance, "UnpublishNow"));
@@ -133,7 +133,7 @@ namespace Workflow.Processes
                 Instance.CompletedDate = DateTime.Now;
                 _instancesService.UpdateInstance(Instance);
 
-                _notifications.Send(Instance, EmailType.ApprovedAndCompletedForScheduler);
+                _emailer.Send(Instance, EmailType.ApprovedAndCompletedForScheduler);
                 // Unpublish will occur via scheduler.
                 Completed?.Invoke(this, new InstanceEventArgs(Instance, "UnpublishAt"));
