+++ Workflow/Workflow.csproj
@@ -40,6 +40,9 @@
     <Compile Include="Api\ExportController.cs" />
     <Content Include="App_Plugins\Workflow\Backoffice\controllers\dashboards\importexport.dashboard.controller.js" />
     <Content Include="App_Plugins\Workflow\Backoffice\controllers\workflow.contenttypeflow.controller.js" />
+    <Content Include="App_Plugins\Workflow\Backoffice\lib\signalr\jquery.signalr-2.2.1.min.js" />
+    <Content Include="App_Plugins\Workflow\Backoffice\notifications.js" />
+    <Content Include="App_Plugins\Workflow\Backoffice\resources\hub.resource.js" />
     <Content Include="App_Plugins\Workflow\Backoffice\views\dialogs\workflow.contenttypeflow.overlay.html" />
     <Content Include="App_Plugins\Workflow\Backoffice\directives\workflow.instance.directive.js" />
     <Content Include="App_Plugins\Workflow\Backoffice\directives\workflow.progressstep.directive.js" />
@@ -86,6 +89,7 @@
     <Compile Include="Services\SettingsService.cs" />
     <Compile Include="Services\TasksService.cs" />
     <Compile Include="Startup\CreateTables.cs" />
+    <Compile Include="Notifications\PlumberHub.cs" />
     <Compile Include="Startup\StringConstraint.cs" />
     <Compile Include="Startup\UmbracoStartup.cs" />
     <Compile Include="Startup\WebApiConfig.cs" />
@@ -100,7 +104,7 @@
     <Compile Include="Models\SettingsPoco.cs" />
     <Compile Include="Repositories\PocoRepository.cs" />
     <Compile Include="Helpers\SqlHelpers.cs" />
-    <Compile Include="Notifications.cs" />
+    <Compile Include="Notifications\Emailer.cs" />
     <Compile Include="Models\SettingsModel.cs" />
     <Compile Include="Processes\DocumentPublishProcess.cs" />
     <Compile Include="Helpers\Utility.cs" />
